plugin: Remove PluginLibrary class from public header
`PluginLibrary` does not have to be public. Resolves an issue with
the included private header file in the public header file.
